pyproject.toml
[project]
name = "linebotfood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "async>=0.6.2",
    "fastapi>=0.124.4",
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "mcp-server-fetch>=2025.1.17",
    "pandas>=2.3.3",
    "pydantic-ai[examples]>=1.34.0",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
